Join when starting a server span with a parent
This replicates the same behavior as exists in zipkin-go-opentracing.
We support adding "span.kind" before a span is started. By doing so, we
can properly decide whether to join a span or start a remote child.

This rewrites the unit tests so that we can ensure dependency links
work, too. The tests were taken from #27.

Closes #27